+++ screens/StackScreens/My/Inquiry.js
@@ -1,7 +1,9 @@
 import React, {useState} from "react";
 import { Text, View, Modal} from "react-native";
 import styled from "styled-components/native";
+import {launchCamera, launchImageLibrary} from 'react-native-image-picker';
 
+import UploadModeModal from "../../../components/UploadModeModal";
 import TopBar from "../../../components/TopBar";
 import BlackButton from "../../../components/BlackButton";
 import BlueButton from "../../../components/BlueButton";
@@ -16,13 +18,13 @@ const Container = styled.ScrollView.attrs(() => ({
     }
 }))`
     flex:1;
-    margin-top: 10px;
-    margin-bottom: 10px;
+    margin: 10px;
+
 `;
 
 const InputBox = styled.View`
     //border: 1px;
-    width: 86%;
+    width: 100%;
     align-self: center;
 `;
     const Text1 = styled.Text`
@@ -51,13 +53,35 @@ const InputBox = styled.View`
         margin-bottom: 16px;
     `;
     const Btn1Box = styled.View`
-        width: 86%;
+        width: 100%;
         align-self: center;
         margin-bottom: 8.2%;
     `;
 
 
 const Inquiry = ({navigation:{navigate}}) => {
+
+    // 선택 사진 또는 촬영된 사진 정보
+    const onPickImage = (res) => { 
+        if (res.didCancel || !res) {
+        return;
+        }
+        console.log("PickImage", res);
+    }
+    
+    // 카메라 촬영
+    const onLaunchCamera = () => {
+        launchCamera(imagePickerOption, onPickImage);
+    };
+    
+    // 갤러리에서 사진 선택
+    const onLaunchImageLibrary = () => {
+        launchImageLibrary(imagePickerOption, onPickImage);
+    };
+
+  // 안드로이드를 위한 모달 visible 상태값
+  const [pictureVisible, setPictureVisible] = useState(false);
+
 
     const [modalVisible, setModalVisible] = useState(false);
     const [saveTitle, setSaveTitle] = useState("");
@@ -79,50 +103,59 @@ const Inquiry = ({navigation:{navigate}}) => {
     }
 
     return(
-    <Container>
-        <TopBar title="문의하기"/>
-        <InputBox>
-            <Text1>제목</Text1>
-            <InputTItle 
-                placeholder = "  문의할 내용을 간략하게 작성해주세요."
-                onChangeText={(title)=>setSaveTitle(title)}>
-            </InputTItle>
-        </InputBox>
-        <InputBox>
-            <Text1>문의 내용</Text1>
-            <InputContent 
-                placeholder = "  문의할 내용을 자세히 작성해주세요."
-                multiline = {true}
-                onChangeText={(content)=>setSaveContent(content)}
-            >
-            </InputContent>
-        </InputBox>
-        <InputBox>
-            <Text1>관련 파일 첨부</Text1>
-            <InputTItle 
-                placeholder = "  파일을 첨부해주세요."
-                onChangeText={(file)=>setSaveFile(file)}
-            >
-            </InputTItle>
-        </InputBox>
-        <Btn1Box>
-            <BlackButton 
-                title="파일 첨부하기"
-                mbottom={10}
-            ></BlackButton>
-        </Btn1Box>
-            <BlueButton 
-                title="제출하기"
-                mbottom={10}
-                click = {setModalVisible}
-            ></BlueButton>
-        <ModalOneOptions
-            title="문의 사항이 제출되었습니다."
-            yestext="확인"
-            visible={modalVisible}
-            setvisible = {setModalVisible}
-        ></ModalOneOptions>
-    </Container>
+        <>
+        <UploadModeModal 
+                visible={pictureVisible} 
+                onClose={() => setPictureVisible(false)}
+                onLaunchCamera={onLaunchCamera}
+                onLaunchImageLibrary={onLaunchImageLibrary} 
+            />
+        <Container>
+            <TopBar title="문의하기"/>
+            <InputBox>
+                <Text1>제목</Text1>
+                <InputTItle 
+                    placeholder = "  문의할 내용을 간략하게 작성해주세요."
+                    onChangeText={(title)=>setSaveTitle(title)}>
+                </InputTItle>
+            </InputBox>
+            <InputBox>
+                <Text1>문의 내용</Text1>
+                <InputContent 
+                    placeholder = "  문의할 내용을 자세히 작성해주세요."
+                    multiline = {true}
+                    onChangeText={(content)=>setSaveContent(content)}
+                >
+                </InputContent>
+            </InputBox>
+            <InputBox>
+                <Text1>관련 파일 첨부</Text1>
+                <InputTItle 
+                    placeholder = "  파일을 첨부해주세요."
+                    onChangeText={(file)=>setSaveFile(file)}
+                >
+                </InputTItle>
+            </InputBox>
+            <Btn1Box>
+                <BlackButton 
+                    title="파일 첨부하기"
+                    mbottom={10}
+                    actOn={setPictureVisible(true)}
+                ></BlackButton>
+            </Btn1Box>
+                <BlueButton 
+                    title="제출하기"
+                    mbottom={10}
+                    click = {setModalVisible}
+                ></BlueButton>
+            <ModalOneOptions
+                title="문의 사항이 제출되었습니다."
+                yestext="확인"
+                visible={modalVisible}
+                setvisible = {setModalVisible}
+            ></ModalOneOptions>
+        </Container>
+    </>
     )
 }
